fix stale funding exclusion leakage on peer TxUpdate (#1394)
* fix: migrate live-cell exclusions on peer funding tx replacement

Keep funding input exclusions aligned with the latest verified peer TxUpdate so stale local hashes cannot pin wallet cells after abort or timeout.


* test: move funding tx tests to ckb test module

Place funding transaction unit coverage in src/ckb/tests/fuding_tests.rs and wire module exports so test helpers remain available under cfg(test).


---------
